+++ provider/sdk_provider.go
@@ -54,12 +54,12 @@ func NewSdkProvider(version string) func() *schema.Provider {
 				"rediscloud_database_modules": dataSourceRedisCloudDatabaseModules(),
 				// Note: rediscloud_regions is served by the Plugin Framework provider
 				// Note: regiscloud_essentials_plan is served by the Plugin Framework provider
-				"rediscloud_essentials_subscription":                    dataSourceRedisCloudEssentialsSubscription(),
-				"rediscloud_essentials_database":                        dataSourceRedisCloudEssentialsDatabase(),
-				"rediscloud_subscription_peerings":                      dataSourceRedisCloudSubscriptionPeerings(),
-				"rediscloud_private_service_connect":                    dataSourcePrivateServiceConnect(),
-				"rediscloud_private_service_connect_endpoints":          dataSourcePrivateServiceConnectEndpoints(),
-				"rediscloud_active_active_subscription":                 dataSourceRedisCloudActiveActiveSubscription(),
+				"rediscloud_essentials_subscription":           dataSourceRedisCloudEssentialsSubscription(),
+				"rediscloud_essentials_database":               dataSourceRedisCloudEssentialsDatabase(),
+				"rediscloud_subscription_peerings":             dataSourceRedisCloudSubscriptionPeerings(),
+				"rediscloud_private_service_connect":           dataSourcePrivateServiceConnect(),
+				"rediscloud_private_service_connect_endpoints": dataSourcePrivateServiceConnectEndpoints(),
+				// Note: rediscloud_active_active_subscription is served by the Plugin Framework provider
 				"rediscloud_active_active_subscription_regions":         dataSourceRedisCloudActiveActiveSubscriptionRegions(),
 				"rediscloud_private_link":                               privatelink.DataSourcePrivateLink(),
 				"rediscloud_private_link_endpoint_script":               privatelink.DataSourcePrivateLinkEndpointScript(),